Let user choose from a list of track sources.
Refactor the Service class (and all its derivatives) to have two methods:
search(), which returns a (possibly empty) list of potential tracks
save(), which takes a track returned from search() and saves it.

Loop through every Service and call its search() method. Display the list
of tracks to the user and prompt for a selection. Call save() on the
selected service, servicetrack pair.

This removes the need to check if search found the correct result
in every service's save() method. It also opens up more UI possibilities.