Provide the YouTube client secrets file.
According to YouTube's docs, even client side javascript apps are expected
to have a client secret (used for refreshing). The docs also state that
installed applications are assumed to be unable to keep secrets. Yet
installed apps have to have a client secret as part of the authorization
flow.

For these reasons, it's probably OK to publish the client secret used by
this application. Worst case, the published secret stops working, and
authorization reverts to the previous scenario, where every user must
apply for their own secret. Best case, the application is significantly
easier to use.